Use "Instant" class instead of "ZonedDateTime" for auditing
As of Spring Boot 2.0 the "java.time.ZonedDateTime" class is not
supported for JPA auditing (AuditAware) out-of-the-box and required to
register a DateTimeProvider bean. See reported issues DATAJPA-1242 (1)
and GH-10743 (2).

The usage of "ZonedDateTime" for auditing has been advised against
because the time zone is really more of a presentation layer concern.
The use of "LocalDateTime" is problematic as well. It does not represent
a point of time but does so only together with a timezone. The current
implementation of Spring "CurrentDateTimeProvider" uses the JVM default
timezone. If an application runs in a cluster across time zones this
results in values of "LocalDateTime" values that can't be compared anymore.

The recommended class to use is "java.time.Instant" instead because there
is no timezone involved, yet it specifies a point in time. For details
see issue DATACMNS-1243 (3) which added converters for
"Instant" <-> "LocalDateTime".

References:

  (1) https://jira.spring.io/browse/DATAJPA-1242
  (2) https://github.com/spring-projects/spring-boot/issues/10743
  (3) https://jira.spring.io/browse/DATACMNS-1243

GH-6