Remove no longer needed Jackson Hibernate 5 module
The "jackson-datatype-hibernate5" module has been used in earlier
versions to handle eagerly fetched entities when the API used to
recursively serialize objects in collections instead of IDs. Is was
used for the "@JsonManagedReference" and "@JsonBackReference"
annotations (1) added to collections within the entity classes to handle
bidirectional relationships. This was a bad API design (recursive tree
serialization for nested key categories) and has already been changed.
This commit removes the still loaded Jackson module and cleanes up the
Maven dependency.

References:

  (1) http://www.baeldung.com/jackson-bidirectional-relationships-and-infinite-recursion

GH-6