Add Git mail mapping file
Added a .mailmap (1) file to link to in documentations to allow
contributors to send mails regarding security issues. This prevents
unnecessary overhead of updating all documents when new team members
are added and additionally adds the main functionality of the file:
Mapping commits when someone uses a different email address.

References:

 (1) https://git-scm.com/docs/git-shortlog#_mapping_authors

GH-9